Manage permissions in users API

+++ users/api.py
@@ -5,12 +5,16 @@ from rest_framework.pagination import PageNumberPagination
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from users.permissions import UserPermission
 from users.serializers import UserSerializer
 
 
 class UserListAPI(APIView):
 
+    permission_classes = (UserPermission,)
+
     def get(self, request):
+        self.check_permissions(request)
         paginator = PageNumberPagination()
         users = User.objects.all()
         paginator.paginate_queryset(users, request)
@@ -18,6 +22,7 @@ class UserListAPI(APIView):
         return paginator.get_paginated_response(serializer.data)
 
     def post(self, request):
+        self.check_permissions(request)
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -28,13 +33,19 @@ class UserListAPI(APIView):
 
 class UserDetailAPI(APIView):
 
+    permission_classes = (UserPermission,)
+
     def get(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
     def put(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user, request.data)
         if serializer.is_valid():
             serializer.save()
@@ -43,7 +54,8 @@ class UserDetailAPI(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
